pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "therapy-scheduler"
version = "0.1.0"
description = "Multi-resource therapy scheduling with OR-Tools and Hydra."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "ortools>=9.10,<9.12",
  "hydra-core>=1.3,<1.4",
  "omegaconf>=2.3,<3",
  "pydantic>=2.6,<3",
  "openpyxl>=3.1,<4",
]

[project.scripts]
therapy-scheduler = "therapy_scheduler.main:main"

[tool.uv]
package = true
dev-dependencies = ["pytest>=7.4"]

[tool.hatch.build.targets.wheel]
packages = ["therapy_scheduler"]
sources = ["src"]
